kriptografi/Stegovault/requirements.txt
Flask==2.0.1
Pillow==8.2.0
pycryptodome==3.10.1
opencv-python==4.5.3.20210927